+++ video_xblock/mixins.py
@@ -54,3 +54,44 @@ class SettingsMixin(XBlock):
         for key, value in self.settings.items():
             fields_dict.setdefault(key, value)
         return fields_dict
+
+
+class LocationMixin(XBlock):
+    """
+    Provides utility methods to access XBlock's `location`.
+
+    Some runtimes, e.g. workbench, don't provide location, hence stubs.
+    """
+
+    @property
+    def block_id(self):
+        """
+        Facade property for `XBlock.location.block_id`.
+
+        Returns stub value if `location` property is unavailabe. E.g. in workbench runtime.
+        """
+        if hasattr(self, 'location'):
+            return self.location.block_id
+        return 'block_id'
+
+    @property
+    def course_key(self):
+        """
+        Facade property for `XBlock.location.course_key`.
+
+        Returns stub value if `location` property is unavailabe. E.g. in workbench runtime.
+        """
+        if hasattr(self, 'location'):
+            return self.location.course_key
+        return 'course_key'
+
+    @property
+    def deprecated_string(self):
+        """
+        Facade property for `XBlock.location.course_key`.
+
+        Returns stub value if `location` property is unavailabe. E.g. in workbench runtime.
+        """
+        if hasattr(self, 'location'):
+            return self.location.to_deprecated_string()
+        return 'deprecated_string'