+++ video_xblock/utils.py
@@ -3,13 +3,23 @@ Video xblock helpers.
 """
 
 from HTMLParser import HTMLParser
+from importlib import import_module
 import os.path
 import pkg_resources
 
 from django.template import Engine, Context, Template
+from xblockutils.resources import ResourceLoader
 
 
 html_parser = HTMLParser()  # pylint: disable=invalid-name
+loader = ResourceLoader(__name__)  # pylint: disable=invalid-name
+
+
+def import_from(module, klass):
+    """
+    Dynamic equivalent for 'from module import klass'.
+    """
+    return getattr(import_module(module), klass)
 
 
 def resource_string(path):