+++ video_xblock/workbench/__init__.py
@@ -1,0 +1,3 @@
+"""
+Workbench submodule add XBlock-SDK workbench scenarios support.
+"""